A prototype of a auto-configuration mechanism for Bazel
This mechanism allow the creation of modules that can discover
artifacts on the system and generate packages in a "hiddden"
directory that is mapped using the package_path mechanisms.

This is a very preliminary version of the system written outside
of Bazel. It lacks proper tests and documentation and the coding
style might not be correct.